ansible: add etcd system units and service start/stop scripts
The scripts allow for orchestrating etcd service bringup in a cluster i.e. adding
and removing a node's membership from etcd.

etcd service is installed on nodes in the 'service-master' host-group. This
shall allow cluster manger to appropriately configure etcd on a node when it is
commissioned. And remove it from cluster when it is decommissioned.

Also added a few packages viz. lshw and serf in the base image. These are needed
for node discovery.

Signed-off-by: Madhav Puri <madhav.puri@gmail.com>