ansible: fix docker config and make docker service restart conditional
- The docker's environment config in 'etc/default/docker' as set by the `base`
role was not taking effect. It seems the file is not automatically picked by
systemd and is not configured to be used as environment as well. Fixed this by
creating the file under `docker.service.d` directory which is read by systemd.

- the docker 'role' may be called as part of multiple host-group. Making the
service restart conditional ensures that docker is only restarted if things
change, making it less disrutpive.

Signed-off-by: Madhav Puri <madhav.puri@gmail.com>