ansible: etcd provisioning enhancements
This enhancement splits etcd cluster/peer configuration into
two types/scenarios viz. cluster-init and member add.

cluster-init is the scenario when the etcd cluster needs to be initialized
for the first time. This may involve a single node or multiple nodes at once.

member-add scenario is used for subsequent updates to the etcd cluster membership.

This enhancement allows for bringing up a cluster with multiple master nodes in demo
environment.

Signed-off-by: Madhav Puri <madhav.puri@gmail.com>