Prepare for new CI
In addition to adding hooks for the new CI, the version detection
was changed to use git tags directly, instead of having an external
process figure out the git tag, add it to the VERSION file, and
commit/push the change.  Using this method, the version should be
more accurate, and we'll be less likely to clobber a previous
release's artifacts accidentally.  Additionally, this removes the
requirement that the CI be able to commit+push to the git repository.